# perf(toml): Don't recreate tables for make_owned 