--- a/crates/toml/src/map.rs
+++ b/crates/toml/src/map.rs
@@ -277,20 +277,29 @@ where
     /// order.
     ///
     /// Computes in **O(n)** time (average).
+    #[allow(unused_mut)]
     pub(crate) fn retain2<F>(&mut self, mut keep: F)
     where
         F: FnMut(&mut K, &mut V) -> bool,
     {
-        self.map = core::mem::take(&mut self.map)
-            .into_iter()
-            .flat_map(move |(mut k, mut v)| {
-                if keep(&mut k, &mut v) {
-                    Some((k, v))
-                } else {
-                    None
-                }
-            })
-            .collect();
+        #[cfg(feature = "preserve_order")]
+        {
+            use indexmap::map::MutableKeys as _;
+            self.map.retain2(keep);
+        }
+        #[cfg(not(feature = "preserve_order"))]
+        {
+            self.map = core::mem::take(&mut self.map)
+                .into_iter()
+                .flat_map(move |(mut k, mut v)| {
+                    if keep(&mut k, &mut v) {
+                        Some((k, v))
+                    } else {
+                        None
+                    }
+                })
+                .collect();
+        }
     }
 }