Before
```console
$ cargo bench --bench 0-cargo -Ffast_hash -- 0_cargo::toml::detable_owned::2-features
    Finished `bench` profile [optimized] target(s) in 0.02s
     Running benches/0-cargo.rs (/home/epage/src/personal/toml/target/release/deps/0_cargo-ea11bd9a52599c35)
Timer precision: 13 ns
0_cargo              fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ toml                            │               │               │               │         │
   ╰─ detable_owned                │               │               │               │         │
      ╰─ 2-features  334.4 µs      │ 724.1 µs      │ 351.8 µs      │ 359.9 µs      │ 100     │ 100
```
After refactor
```console
$ cargo bench --bench 0-cargo -Ffast_hash -- 0_cargo::toml::detable_owned::2-features
    Finished `bench` profile [optimized] target(s) in 0.02s
     Running benches/0-cargo.rs (/home/epage/src/personal/toml/target/release/deps/0_cargo-ea11bd9a52599c35)
Timer precision: 11 ns
0_cargo              fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ toml                            │               │               │               │         │
   ╰─ detable_owned                │               │               │               │         │
      ╰─ 2-features  353.8 µs      │ 1.108 ms      │ 397.8 µs      │ 431.1 µs      │ 100     │ 100
```
After perf work
```console
$ cargo bench --bench 0-cargo -Ffast_hash -- 0_cargo::toml::detable_owned::2-features
   Compiling toml v0.8.23 (/home/epage/src/personal/toml/crates/toml)
   Compiling toml_benchmarks v0.0.0 (/home/epage/src/personal/toml/crates/benchmarks)
    Finished `bench` profile [optimized] target(s) in 19.01s
     Running benches/0-cargo.rs (/home/epage/src/personal/toml/target/release/deps/0_cargo-ea11bd9a52599c35)
Timer precision: 11 ns
0_cargo              fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ toml                            │               │               │               │         │
   ╰─ detable_owned                │               │               │               │         │
      ╰─ 2-features  305.1 µs      │ 670.4 µs      │ 326.9 µs      │ 334.7 µs      │ 100     │ 100
```

As a point of comparison
```console
$ cargo bench --bench 0-cargo -Ffast_hash -- 0_cargo::toml::detable::2-features
    Finished `bench` profile [optimized] target(s) in 0.02s
     Running benches/0-cargo.rs (/home/epage/src/personal/toml/target/release/deps/0_cargo-ea11bd9a52599c35)
Timer precision: 17 ns
0_cargo              fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ toml                            │               │               │               │         │
   ╰─ detable                      │               │               │               │         │
      ╰─ 2-features  265.8 µs      │ 651 µs        │ 288.9 µs      │ 294.1 µs      │ 100     │ 100
```